-feat: adicionado endpoint para criação de transação Foi implementado o CreateTransaction no Repository,Usecase, Controller e criado o endpoint no main. Testes pendente. -refactor: refatoração de sintaxe no connectionString dentro do config.go

+++ model/transaction.go
@@ -3,7 +3,7 @@ package model
 import "time"
 
 type Transaction struct {
-	ID          int64     `json:"id_transaction"`
+	ID          int       `json:"id_transaction"`
 	Description string    `json:"description"`
 	Date        time.Time `json:"date"`
 	Value       float64   `json:"value"`
